+++ src/DAP/Types.hs
@@ -72,6 +72,8 @@ module DAP.Types
   , PathFormat                         (..)
     -- * Command
   , Command                            (..)
+    -- * Reverse Command
+  , ReverseCommand                     (..)
     -- * Event
   , EventType                          (..)
     -- ** Events
@@ -99,13 +101,14 @@ module DAP.Types
   , AdaptorState                       (..)
   , AdaptorLocal(..)
   , AppStore
-  , MonadIO(..)
+  , MonadIO
     -- * Errors
   , AdaptorException                   (..)
   , ErrorMessage                       (..)
   , ErrorResponse                      (..)
     -- * Request
   , Request                            (..)
+  , ReverseRequestResponse             (..)
     -- * Misc.
   , PayloadSize
   , Seq
@@ -121,6 +124,7 @@ module DAP.Types
   , LoadedSourcesResponse              (..)
   , ModulesResponse                    (..)
   , ReadMemoryResponse                 (..)
+  , RunInTerminalResponse              (..)
   , ScopesResponse                     (..)
   , SetExpressionResponse              (..)
   , SetVariableResponse                (..)
@@ -153,6 +157,8 @@ module DAP.Types
   , RestartArguments                   (..)
   , RestartFrameArguments              (..)
   , ReverseContinueArguments           (..)
+  , RunInTerminalRequestArguments      (..)
+  , RunInTerminalRequestArgumentsKind  (..)
   , ScopesArguments                    (..)
   , SetBreakpointsArguments            (..)
   , SetDataBreakpointsArguments        (..)
@@ -172,7 +178,6 @@ module DAP.Types
   , ThreadsArguments                   (..)
   , VariablesArguments                 (..)
   , WriteMemoryArguments               (..)
-  , RunInTerminalResponse              (..)
     -- * defaults
   , defaultBreakpoint
   , defaultBreakpointLocation
@@ -219,7 +224,7 @@ import           Data.Aeson                      ( (.:), (.:?), withObject, with
                                                  , FromJSON(parseJSON), Value, KeyValue((.=))
                                                  , ToJSON(toJSON), genericParseJSON, defaultOptions
                                                  )
-import           Data.Aeson.Types                ( Pair, typeMismatch )
+import           Data.Aeson.Types                ( Pair, typeMismatch, Parser )
 import           Data.Proxy                      ( Proxy(Proxy) )
 import           Data.String                     ( IsString(..) )
 import           Data.Time                       ( UTCTime )
@@ -361,10 +366,35 @@ data Request
 ----------------------------------------------------------------------------
 instance FromJSON Request where
   parseJSON = withObject "Request" $ \o -> do
+    "request" <- (o .: "type") :: Parser String
     Request
       <$> o .:? "arguments"
       <*> o .: "seq"
       <*> o .: "command"
+----------------------------------------------------------------------------
+data ReverseRequestResponse
+  = ReverseRequestResponse
+  { body :: Maybe Value
+    -- ^ Request arguments
+    --
+  , reverseRequestResponseSeqNum :: Seq
+    -- ^ Request sequence number
+    --
+  , reverseRequestCommand :: ReverseCommand
+    -- ^ Command of Request
+    --
+  , success :: Bool
+    -- ^ Whether the reverse request was successful
+  } deriving stock (Show)
+----------------------------------------------------------------------------
+instance FromJSON ReverseRequestResponse where
+  parseJSON = withObject "ReverseRequestResponse" $ \o -> do
+    "response" <- (o .: "type") :: Parser String
+    ReverseRequestResponse
+      <$> o .:? "body"
+      <*> o .: "seq"
+      <*> o .: "command"
+      <*> o .: "success"
 ----------------------------------------------------------------------------
 data Breakpoint
   = Breakpoint
@@ -894,8 +924,6 @@ instance ToJSON EventType where
 ----------------------------------------------------------------------------
 data Command
   = CommandCancel
-  | CommandRunInTerminal
-  | CommandStartDebugging
   | CommandInitialize
   | CommandConfigurationDone
   | CommandLaunch
@@ -952,6 +980,24 @@ instance FromJSON Command where
 ----------------------------------------------------------------------------
 instance ToJSON Command where
   toJSON (CustomCommand x) = toJSON x
+  toJSON cmd = genericToJSONWithModifier cmd
+----------------------------------------------------------------------------
+data ReverseCommand
+  = ReverseCommandRunInTerminal
+  | ReverseCommandStartDebugging
+  deriving stock (Show, Eq, Read, Generic)
+----------------------------------------------------------------------------
+instance FromJSON ReverseCommand where
+  parseJSON = withText name $ \command ->
+    case readMaybe (name <> capitalize (T.unpack command)) of
+      Just cmd ->
+        pure cmd
+      Nothing ->
+        fail $ "Unknown reverse command: " ++ show command
+    where
+      name = show (typeRep (Proxy @ReverseCommand))
+----------------------------------------------------------------------------
+instance ToJSON ReverseCommand where
   toJSON cmd = genericToJSONWithModifier cmd
 ----------------------------------------------------------------------------
 data ErrorMessage
@@ -1095,6 +1141,8 @@ data RunInTerminalResponse
 ----------------------------------------------------------------------------
 instance ToJSON RunInTerminalResponse where
   toJSON = genericToJSONWithModifier
+instance FromJSON RunInTerminalResponse where
+  parseJSON = genericParseJSONWithModifier
 ----------------------------------------------------------------------------
 data ModulesResponse
   = ModulesResponse
@@ -2688,6 +2736,9 @@ data RunInTerminalRequestArgumentsKind
   | RunInTerminalRequestArgumentsKindExternal
   deriving stock (Show, Eq, Generic)
 ----------------------------------------------------------------------------
+instance ToJSON RunInTerminalRequestArgumentsKind where
+  toJSON = genericToJSONWithModifier
+----------------------------------------------------------------------------
 instance FromJSON RunInTerminalRequestArgumentsKind where
   parseJSON = genericParseJSONWithModifier
 ----------------------------------------------------------------------------
@@ -2727,6 +2778,9 @@ data RunInTerminalRequestArguments
     -- special characters may not be portable across shells.
     --
   } deriving stock (Show, Eq, Generic)
+----------------------------------------------------------------------------
+instance ToJSON RunInTerminalRequestArguments where
+  toJSON = genericToJSONWithModifier
 ----------------------------------------------------------------------------
 instance FromJSON RunInTerminalRequestArguments where
   parseJSON = genericParseJSONWithModifier